fix guard check, add more tests
