+++ packages/browser-integration-tests/suites/integrations/ContextLines/inline/template.html
@@ -5,7 +5,6 @@
   </head>
   <body>
       <button id="inline-error-btn" onclick="throw new Error('Error with context lines')">Click me</button>
-      <button id="script-error-btn">Click me too</button>
   </body>
   <footer>
     Some text...
