+++ packages/browser-integration-tests/suites/integrations/ContextLines/noAddedLines/template.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+  </head>
+  <body>
+      <button id="script-error-btn">Click me</button>
+  </body>
+  <footer>
+    Some text...
+  </foot>
+</html>
