+++ packages/browser-integration-tests/suites/integrations/ContextLines/scriptTag/template.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <script>
+      function throwTestError() {
+        throw new Error('Error with context lines')
+      }
+    </script>
+  </head>
+  <body>
+      <button id="inline-error-btn" onclick="throwTestError()">Click me</button>
+  </body>
+  <footer>
+    Some text...
+  </foot>
+</html>
